#  Faces: Improve performance when manual tagging