--- a/internal/api/markers.go
+++ b/internal/api/markers.go
@@ -228,7 +228,7 @@ func UpdateMarker(router *gin.RouterGroup) {
 			return
 		} else if changed {
 			if marker.FaceID != "" && marker.SubjUID != "" && marker.SubjSrc == entity.SrcManual {
-				if res, err := get.Faces().Optimize(); err != nil {
+				if res, err := get.Faces().OptimizeFor(marker.SubjUID); err != nil {
 					log.Errorf("faces: %s (optimize)", err)
 				} else if res.Merged > 0 {
 					log.Infof("faces: merged %s", english.Plural(res.Merged, "cluster", "clusters"))