# (1) Please provide a concise description of your pull request
## What does it implement / fix / improve? Why?

**What?**: Per [#3124](https://github.com/photoprism/photoprism/issues/3124), this PR implements strategy [(1)](https://github.com/photoprism/photoprism/issues/3124#issuecomment-2558299360) to improve `Optimize()` performance by returning the correct number of rows modified in `PurgeOrphanFaces` and  treating the result `0` as an error condition in `MergeFaces`. Logging around this has been improved in `faces.go` and `faces_optimize.go`. 

**Why?**: To improve performance when manually tagging faces, in the case that it becomes necessary to call `Optimize()`. Please see [issue 3124](https://github.com/photoprism/photoprism/issues/3124#issuecomment-2558299360) for the detailed description of the problem. 

## Are the changes related to an existing issue?

Yes: [#3124](https://github.com/photoprism/photoprism/issues/3124)

# (2) CLA (see https://www.photoprism.app/cla)

Signed and scanned

# (3) Acceptance Criteria:

- [x] Features and enhancements must be fully implemented so that they can be released at any time without additional work
- [ ] Automated unit and/or acceptance tests are mandatory to ensure the changes work as expected and to reduce repetitive manual work *see below!*
- [x] Frontend components must be responsive to work and look properly on phones, tablets, and desktop computers; you must have tested them on all major browsers and different devices *- no frontend changes, but did test on chrome, firefox, safari*
- [x] Documentation and translation updates should be provided if needed *no translations modified*
- [x] In case you submit database-related changes, they must be tested and compatible with SQLite 3 and MariaDB 10.5.12+ *no database schema changes*

<!--

Since reviewing, testing and finally merging pull requests requires significant resources on our side, this can take several months if it's not just a small fix, especially if extensive testing is required to prevent bugs from getting into our stable version.

We thank you for your patience! :)

-->
# Testing issues 

Development machine is Windows 10 (Docker WSLv2). Followed instructions, having test errors that seem disconnected from the modified code:

```
FAIL | UpdateMarker (0.08s)
FAIL |   UpdateMarker/update_cluster_with_existing_subject_2 (0.02s)
     |         markers_test.go:272:
     |                  Error Trace:    /go/src/github.com/photoprism/photoprism/internal/api/markers_test.go:272
     |                  Error:          Not equal:
     |                                  expected: 200
     |                                  actual  : 500
     |                  Test:           TestUpdateMarker/update_cluster_with_existing_subject_2
...
     | time="2024-12-30T18:39:51Z" level=debug msg="api-v1: abort /api/v1/zip with code 400 (unable to do that)"
     | time="2024-12-30T18:39:51Z" level=error msg="download: /go/src/github.com/photoprism/photoprism/storage/testdata/temp/zip/xxx not found"
FAIL |  github.com/photoprism/photoprism/internal/api   72.942s
...
FAIL | MediaFile_NeedsExifToolJson (0.12s)
FAIL |   MediaFile_NeedsExifToolJson/false (0.04s)
     |         mediafile_meta_test.go:84:
     |                  Error Trace:    /go/src/github.com/photoprism/photoprism/internal/photoprism/mediafile_meta_test.go:84
     |                  Error:          Should be true
     |                  Test:           TestMediaFile_NeedsExifToolJson/false
FAIL |  github.com/photoprism/photoprism/internal/photoprism
```

FWIW, in my environment, these tests **all fail on a completely vanilla develop branch**. During my attempt at implementation, there were other errors/failing tests that *were* caught and are now passing, so I assume that these test failures are alright. 

In lieu of passing tests, I built using `make preview` (changing the docker-buildx.sh script to use `--load` instead of `--push`) and set the image in my main photoprism deployment (after backing everything up) to use the local image. Observations will follow in a comment below.